rewrite contracts addresses hooks
