+++ src/hooks/dao/index.ts
@@ -1,14 +1,13 @@
 export * from "./proposals"
 export { default as useCreateDAO } from "./useCreateDAO"
 export { default as useGovSettingsNewSettingId } from "./useGovSettingsNewSettingId"
-export {
-  default as useGovPoolSetting,
-  useGovSettingsAddress,
-} from "./useGovPoolSetting"
+export { default as useGovPoolSetting } from "./useGovPoolSetting"
 export * from "./useGovPool"
 export { default as useGovPoolDeposit } from "./useGovPoolDeposit"
 export { default as useGovPoolValidators } from "./useGovPoolValidators"
-export * from "./useGovValidatorsContractAddress"
+export { useGovSettingsContractAddress } from "./useGovSettingsContractAddress"
+export { useGovValidatorsContractAddress } from "./useGovValidatorsContractAddress"
+export { useGovUserKeeperContractAddress } from "./useGovUserKeeperContractAddress"
 export { default as useGovBalance } from "./useGovBalance"
 export { default as useGovValidatorsInternalSettings } from "./useGovValidatorsInternalSettings"
 export { default as useGovValidatorsTokenTotalSupply } from "./useGovValidatorsTokenTotalSupply"
