+++ src/hooks/dao/useGovPoolSetting.ts
@@ -59,26 +59,4 @@ const useGovPoolSetting = ({
   return [result, loading, error]
 }
 
-export const useGovSettingsAddress = (daoAddress: string) => {
-  const govPoolContract = useGovPoolContract(daoAddress)
-
-  const [govSettingsAddress, setGovSettingsAddress] = useState<string>("")
-
-  useEffect(() => {
-    const setupGovSettingsAddress = async () => {
-      if (!govPoolContract) return
-
-      try {
-        const { settings } = await govPoolContract.getHelperContracts()
-        setGovSettingsAddress(settings)
-      } catch (error) {
-        console.log(error)
-      }
-    }
-    setupGovSettingsAddress()
-  }, [govPoolContract])
-
-  return govSettingsAddress
-}
-
 export default useGovPoolSetting
